@@ -14,8 +14,8 @@
     <description>ccd-application</description>
     <properties>
         <default-package>ca.bc.gov.open.ccd</default-package>
-        <java.version>11</java.version>
-        <log4j2.version>2.17.1</log4j2.version>
+        <maven.compiler.source>17</maven.compiler.source>
+        <maven.compiler.target>17</maven.compiler.target>
     </properties>
 
     <repositories>
