@@ -6,7 +6,7 @@
     <parent>
         <groupId>org.springframework.boot</groupId>
         <artifactId>spring-boot-starter-parent</artifactId>
-        <version>2.7.13</version>
+        <version>3.1.2</version>
     </parent>
     <groupId>ca.bc.gov.open</groupId>
     <artifactId>jag-ccd</artifactId>
@@ -14,7 +14,7 @@
     <name>jag-ccd</name>
     <description>Court Clerk Display</description>
     <properties>
-        <java.version>11</java.version>
+        <java.version>17</java.version>
         <log4j2.version>2.17.1</log4j2.version>
         <snakeyaml.version>2.0</snakeyaml.version>
     </properties>
@@ -32,7 +32,6 @@
         <module>ccd-court-list-models</module>
         <module>ccd-civil-models</module>
         <module>ccd-common-models</module>
-        <module>AutomatedTests</module>
         <module>comparison-tool-1.1</module>
     </modules>
 
@@ -65,13 +64,6 @@
                             </indent>
                         </format>
                     </formats>
-                    <!-- define a language-specific format -->
-                    <java>
-                        <googleJavaFormat>
-                            <version>1.8</version>
-                            <style>AOSP</style>
-                        </googleJavaFormat>
-                    </java>
                 </configuration>
             </plugin>
             <plugin>
