@@ -64,7 +64,7 @@ public class CourtControllerTests {
         is.setStatuteActCd("A");
         is.setStatuteSectionCd("A");
         is.setStatuteSectionDsc("A");
-        fi.setIssue(Collections.singletonList(is));
+        fi.getIssue().add(is);
 
         crl.setFileInformation(fi);
 
@@ -104,7 +104,7 @@ public class CourtControllerTests {
         at.setPhoneNumber("A");
         at.setInstruction("A");
         at.setOtherRoleName("A");
-        crl.setAttendanceMethod(Collections.singletonList(at));
+        crl.getAttendanceMethod().add(at);
 
         var ac = new AppearanceCount();
         ac.setAppearanceCountId("A");
@@ -127,7 +127,7 @@ public class CourtControllerTests {
         ac.setFindingCode("A");
         ac.setFindingDate(Instant.now());
 
-        crl.setAppearanceCount(Collections.singletonList(ac));
+        crl.getAppearanceCount().add(ac);
 
         var bn = new Bans();
         bn.setBanTypeCd("A");
@@ -138,7 +138,7 @@ public class CourtControllerTests {
         bn.setBanStatuteId("A");
         bn.setBanCommentText("A");
         bn.setBanAcprId("A");
-        crl.setBans(Collections.singletonList(bn));
+        crl.getBans().add(bn);
 
         var bov = new BailOrderToVary();
         bov.setFormTypeCd("A");
@@ -147,7 +147,7 @@ public class CourtControllerTests {
         bov.setDocmIssueDate(Instant.now());
         bov.setDocmImageId("A");
         bov.setDocmStatus("A");
-        crl.setBailOrderToVary(Collections.singletonList(bov));
+        crl.getBailOrderToVary().add(bov);
 
         var sov = new SentOrderToVary();
         sov.setFormTypeCd("A");
@@ -156,7 +156,7 @@ public class CourtControllerTests {
         sov.setDocmIssueDate(Instant.now());
         sov.setDocmImageId("A");
         sov.setDocmStatus("A");
-        crl.setSentOrderToVary(Collections.singletonList(sov));
+        crl.getSentOrderToVary().add(sov);
 
         var an = new AgeNotice();
         an.setEventDate(Instant.now());
@@ -167,7 +167,7 @@ public class CourtControllerTests {
         an.setProvenBy("A");
         an.setNoticeTo("A");
 
-        crl.setAgeNotice(Collections.singletonList(an));
+        crl.getAgeNotice().add(an);
 
         var st = new SpeakerType();
         st.setSpeakerId("A");
@@ -182,19 +182,19 @@ public class CourtControllerTests {
         se.setSpeakerEventTime("A");
         se.setSpeakerEventText("A");
 
-        st.setSpeakerEvent(Collections.singletonList(se));
-        crl.setSpeaker(Collections.singletonList(st));
+        st.getSpeakerEvent().add(se);
+        crl.getSpeaker().add(st);
 
         ArrestWarrant aw = new ArrestWarrant();
         aw.setFileNumberText("A");
         aw.setWarrantDate(Instant.now());
-        crl.setArrestWarrant(Collections.singletonList(aw));
+        crl.getArrestWarrant().add(aw);
 
         var crp = new CRProtectionOrderType();
         crp.setPORConditionText("A");
         crp.setOrderTypeDsc("A");
         crp.setPOROrderIssueDate(Instant.now());
-        crl.setProtectionOrder(Collections.singletonList(crp));
+        crl.getProtectionOrder().add(crp);
 
         var sa = new ScheduledAppearance();
         sa.setAppearanceId("A");
@@ -211,15 +211,15 @@ public class CourtControllerTests {
         hr.setJudgeName("A");
         hr.setHearingRestrictionDate(Instant.now());
 
-        crl.setHearingRestriction(Collections.singletonList(hr));
-        crl.setScheduledAppearance(Collections.singletonList(sa));
+        crl.getHearingRestriction().add(hr);
+        crl.getScheduledAppearance().add(sa);
 
         var co = new CFCOrderType();
         co.setCFCOrderIssueDate(Instant.now());
         co.setOrderTypeDsc("A");
         co.setCFCConditionText("A");
 
-        crl.setCFCOrder(Collections.singletonList(co));
+        crl.getCFCOrder().add(co);
 
         var crl2 = new CivilCourtListType();
         crl2.setAppearanceId("A");
@@ -259,7 +259,7 @@ public class CourtControllerTests {
 
         AssetType as = new AssetType();
         as.setAssetTypeDescription("A");
-        crl2.setAsset(Collections.singletonList(as));
+        crl2.getAsset().add(as);
 
         DocumentType dt = new DocumentType();
         dt.setAppearanceID("A");
@@ -285,13 +285,13 @@ public class CourtControllerTests {
         is2.setIssueDescription("A");
         is2.setIssueType("A");
         is2.setIssueNumber("A");
-        dt.setIssue(Collections.singletonList(is2));
+        dt.getIssue().add(is2);
 
         var ft = new FiledByType();
         ft.setFiledByName("A");
         ft.setRoleTypeCode("A");
-        dt.setFiledBy(Collections.singletonList(ft));
-        crl2.setDocument(Collections.singletonList(dt));
+        dt.getFiledBy().add(ft);
+        crl2.getDocument().add(dt);
 
         var pt = new PartyType();
         pt.setPartyId("A");
@@ -312,25 +312,25 @@ public class CourtControllerTests {
         ct.setPhoneNumber("A");
         ct.setCounselId("A");
 
-        pt.setCounsel(Collections.singletonList(ct));
+        pt.getCounsel().add(ct);
 
         var rt = new RepresentativeType();
         rt.setInstruction("A");
         rt.setAttendanceMethodCd("A");
         rt.setPhoneNumber("A");
         rt.setRepFullName("A");
-        pt.setRepresentative(Collections.singletonList(rt));
+        pt.getRepresentative().add(rt);
 
         var lrt = new LegalRepresentativeType();
         lrt.setLegalRepFullName("A");
         lrt.setLegalRepTypeDsc("A");
-        pt.setLegalRepresentative(Collections.singletonList(lrt));
+        pt.getLegalRepresentative().add(lrt);
 
         var pr2 = new PartyRoleType2();
         pr2.setRoleTypeCd("A");
         pr2.setRoleTypeDsc("A");
 
-        pt.setPartyRole(Collections.singletonList(pr2));
+        pt.getPartyRole().add(pr2);
 
         PartyNameType pnt = new PartyNameType();
         pnt.setFirstGivenNm("A");
@@ -341,9 +341,9 @@ public class CourtControllerTests {
         pnt.setSecondGivenNm("A");
         pnt.setThirdGivenNm("A");
         pnt.setOrganizationNm("A");
-        pt.setPartyName(Collections.singletonList(pnt));
+        pt.getPartyName().add(pnt);
 
-        crl2.setParties(Collections.singletonList(pt));
+        crl2.getParties().add(pt);
 
         SpeakerType st2 = new SpeakerType();
         st2.setSpeakerId("A");
@@ -357,7 +357,7 @@ public class CourtControllerTests {
         se2.setSpeakerEventText("A");
         se2.setSpeakerEventDate("A");
         se2.setSpeakerEventTime("A");
-        st2.setSpeakerEvent(Collections.singletonList(se2));
+        st2.getSpeakerEvent().add(se2);
 
         var pot = new ProtectionOrderType();
         pot.setPOROrderIssueDate(Instant.now());
@@ -368,10 +368,10 @@ public class CourtControllerTests {
         rpt.setRestrainingPartyName("A");
         ProtectedPartyNameType ppt = new ProtectedPartyNameType();
         ppt.setProtectedPartyName("A");
-        pot.setRestrainingPartyName(Collections.singletonList(rpt));
-        pot.setProtectedPartyName(Collections.singletonList(ppt));
+        pot.getRestrainingPartyName().add(rpt);
+        pot.getProtectedPartyName().add(ppt);
 
-        crl2.setProtectionOrder(Collections.singletonList(pot));
+        crl2.getProtectionOrder().add(pot);
 
         var sat = new ScheduledAppearanceType();
         sat.setAppearanceId("A");
@@ -383,19 +383,19 @@ public class CourtControllerTests {
         sat.setEstDurationHours("A");
         sat.setEstDurationMins("A");
 
-        crl2.setScheduledAppearance(Collections.singletonList(sat));
+        crl2.getScheduledAppearance().add(sat);
 
         var hr2 = new HearingRestrictionType();
         hr2.setHearingRestrictionDate(Instant.now());
         hr2.setJudgeName("A");
         hr2.setHearingRestrictionDate(Instant.now());
-        crl2.setHearingRestriction(Collections.singletonList(hr2));
+        crl2.getHearingRestriction().add(hr2);
 
         var caw = new CivilArrestWarrantType();
         caw.setWarrantDate(Instant.now());
         caw.setWarrantTypeCd("A");
         caw.setWarrantTypeDsc("A");
-        crl2.setArrestWarrant(Collections.singletonList(caw));
+        crl2.getArrestWarrant().add(caw);
 
         var ud = new UnscheduledDocumentType();
         ud.setDocumentId("A");
@@ -404,9 +404,9 @@ public class CourtControllerTests {
         FiledByType fb = new FiledByType();
         fb.setRoleTypeCode("A");
         fb.setRoleTypeCode("A");
-        ud.setFiledBy(Collections.singletonList(fb));
+        ud.getFiledBy().add(fb);
 
-        crl2.setUnscheduledDocument(Collections.singletonList(ud));
+        crl2.getUnscheduledDocument().add(ud);
 
         var ov = new OrderToVaryType();
         ov.setAdjudicatorName("A");
@@ -414,7 +414,7 @@ public class CourtControllerTests {
         ov.setDocumentId("A");
         ov.setDocumentTypeDsc("A");
 
-        crl2.setOrderToVary(Collections.singletonList(ov));
+        crl2.getOrderToVary().add(ov);
 
         var courtListTypes = new CourtListTypes();
         courtListTypes.setCourtLocationName("A");
@@ -424,8 +424,8 @@ public class CourtControllerTests {
         fs.setFileNumber("A");
         fs.setCourtDivisionCd("A");
         courtListTypes.setFileSearchParameter(fs);
-        courtListTypes.setCivilCourtList(Collections.singletonList(crl2));
-        courtListTypes.setCriminalCourtList(Collections.singletonList(crl));
+        courtListTypes.getCivilCourtList().add(crl2);
+        courtListTypes.getCriminalCourtList().add(crl);
 
         var out = new GetCrtListResponse();
         out.setCourtLists(courtListTypes);
