@@ -55,8 +55,8 @@ public class ProcessControllerTests {
         af.setCountId("A");
         af.setCountNumber("A");
 
-        vdt.setApplyToFile(Collections.singletonList(af));
-        var.setVariationDetail(Collections.singletonList(vdt));
+        vdt.getApplyToFile().add(af);
+        var.getVariationDetail().add(vdt);
 
         var out = new ProcessVariationResponse();
         out.setStatus("A");
@@ -89,7 +89,7 @@ public class ProcessControllerTests {
         app.setFileNumber("A");
         app.setFileCourtDivisionCd("A");
 
-        sp.setApplyToFile(Collections.singletonList(app));
+        sp.getApplyToFile().add(app);
 
         var sd = new SpeakerDetailsType();
         sd.setSpeakerId("A");
@@ -104,7 +104,7 @@ public class ProcessControllerTests {
         se.setSpeakerEventTime("A");
         se.setSpeakerEventText("A");
 
-        sd.setSpeakerEvent(Collections.singletonList(se));
+        sd.getSpeakerEvent().add(se);
 
         req.setSpeaker(sp);
 
@@ -172,7 +172,7 @@ public class ProcessControllerTests {
         af.setFileNumber("A");
         af.setFileSeqNo("A");
 
-        ni.setApplyToFile(Collections.singletonList(af));
+        ni.getApplyToFile().add(af);
 
         PartyType pt = new PartyType();
         pt.setPartyId("A");
@@ -181,7 +181,7 @@ public class ProcessControllerTests {
         pt.setPhoneNumber("A");
         pt.setInstruction("A");
 
-        ni.setParty(Collections.singletonList(pt));
+        ni.getParty().add(pt);
 
         var nd = new NewDocumentFiledType();
         nd.setAppearanceId("A");
@@ -216,17 +216,17 @@ public class ProcessControllerTests {
         at.setDocumentId("A");
         at.setFileNumber("A");
 
-        hr.setApplyToFile(Collections.singletonList(at));
+        hr.getApplyToFile().add(at);
 
-        cr.setJudgeComment(Collections.singletonList(jc));
+        cr.getJudgeComment().add(jc);
 
-        cr.setOralApplication(Collections.singletonList(or));
+        cr.getOralApplication().add(or);
 
-        cr.setNewDocumentFiled(Collections.singletonList(nd));
+        cr.getNewDocumentFiled().add(nd);
 
-        cr.setNextAppearance(Collections.singletonList(ni));
-        cd.setIssues(Collections.singletonList(is));
-        cr.setCivilResultDetail(Collections.singletonList(cd));
+        cr.getNextAppearance().add(ni);
+        cd.getIssues().add(is);
+        cr.getCivilResultDetail().add(cd);
 
         req.setCivilResult(cr);
 
@@ -270,7 +270,7 @@ public class ProcessControllerTests {
         var af = new ApplyToFileAppearanceMethodType();
         af.setAppearanceId("A");
         af.setFileNumber("A");
-        at.setApplyToFileAppearanceMethod(Collections.singletonList(af));
+        at.getApplyToFileAppearanceMethod().add(af);
         var pa = new PartyAppearanceMethodType();
         pa.setPartyName("A");
         pa.setPartyRole("A");
@@ -283,8 +283,8 @@ public class ProcessControllerTests {
         pa.setRemoteAgencyIdentifierId("A");
         pa.setRemoteCourtRoomAgencyIdentifierId("A");
         pa.setRemoteCourtRoomCd("A");
-        at.setPartyAppearanceMethod(Collections.singletonList(pa));
-        one.setAppearanceMethodDetail(Collections.singletonList(at));
+        at.getPartyAppearanceMethod().add(pa);
+        one.getAppearanceMethodDetail().add(at);
 
         req.setAppearanceMethod(one);
 
@@ -335,8 +335,8 @@ public class ProcessControllerTests {
         ap.setFileNumber("A");
         ap.setCountNumber("A");
 
-        pd.setApplyToFile(Collections.singletonList(ap));
-        one.setPleaDetail(Collections.singletonList(pd));
+        pd.getApplyToFile().add(ap);
+        one.getPleaDetail().add(pd);
         req.setPlea(one);
 
         var out = new ProcessPleaResponse();
@@ -381,8 +381,8 @@ public class ProcessControllerTests {
         ad.setFileNumber("A");
         ad.setCountId("A");
         ad.setCountNumber("A");
-        ed.setApplyToFileElection(Collections.singletonList(ad));
-        one.setElectionDetails(Collections.singletonList(ed));
+        ed.getApplyToFileElection().add(ad);
+        one.getElectionDetails().add(ed);
 
         req.setElection(one);
         var out = new ProcessElectionResponse();
@@ -425,7 +425,7 @@ public class ProcessControllerTests {
         bt.setRecogTypeCd("A");
         bt.setRecogAmount("A");
         bt.setRecogDeposit("A");
-        bt.setAlternateBailText(Collections.singletonList("A"));
+        bt.getAlternateBailText().add("A");
         bt.setBdpcYN("A");
         bt.setDetainedUnderSec524YN("A");
         bt.setSuretyQty("A");
@@ -436,15 +436,15 @@ public class ProcessControllerTests {
         st.setSuretyAddress("A");
         st.setSuretyDeclaration("A");
         st.setSuretyDeposit("A");
-        bt.setSurety(Collections.singletonList(st));
+        bt.getSurety().add(st);
         var af = new ApplyToFileBailType();
         af.setAppearanceId("A");
         af.setCountId("A");
         af.setCountNumber("A");
         af.setFileNumber("A");
-        bt.setApplyToFile(Collections.singletonList(af));
+        bt.getApplyToFile().add(af);
 
-        one.setBailDetail(Collections.singletonList(bt));
+        one.getBailDetail().add(bt);
         req.setBailDocInput(one);
 
         var out = new ProcessBailResponse();
@@ -513,8 +513,8 @@ public class ProcessControllerTests {
         nct.setPhoneNumber("A");
         nct.setInstruction("A");
 
-        nc.setNextAppMethod(Collections.singletonList(nct));
-        cd.setApplyToFile(Collections.singletonList(af));
+        nc.getNextAppMethod().add(nct);
+        cd.getApplyToFile().add(af);
 
         var gd = new GenerateDocumentCRType();
         gd.setFormTypeCd("A");
@@ -524,8 +524,8 @@ public class ProcessControllerTests {
         gd.setReferralReason("A");
         gd.setReferralComment("A");
 
-        cd.setGenerateDocument(Collections.singletonList(gd));
-        one.setCriminalResultsDetail(Collections.singletonList(cd));
+        cd.getGenerateDocument().add(gd);
+        one.getCriminalResultsDetail().add(cd);
         req.setCriminalResult(one);
 
         var out = new ProcessCriminalResultResponse();
@@ -568,7 +568,7 @@ public class ProcessControllerTests {
         af.setAppearanceId("A");
         af.setFileNumber("A");
 
-        one.setAgeNoticeDetail(Collections.singletonList(an));
+        one.getAgeNoticeDetail().add(an);
 
         var ag = new AgeNoticeEventType();
         ag.setEventType("A");
@@ -578,8 +578,8 @@ public class ProcessControllerTests {
         ag.setProvenBy("A");
         ag.setNoticeTo("A");
 
-        an.setAgeNoticeEvent(Collections.singletonList(ag));
-        an.setApplyToFileAgeNotice(Collections.singletonList(af));
+        an.getAgeNoticeEvent().add(ag);
+        an.getApplyToFileAgeNotice().add(af);
 
         req.setAgeNotice(one);
 
@@ -628,8 +628,8 @@ public class ProcessControllerTests {
         cf.setCivilPhysicalFileId("A");
         cf.setFileNumber("A");
         cf.setFileCourtDivisionCd("A");
-        mc.setCalledFile(Collections.singletonList(cf));
-        one.setMatterCallDetails(Collections.singletonList(mc));
+        mc.getCalledFile().add(cf);
+        one.getMatterCallDetails().add(mc);
 
         req.setMatterCall(one);
 
@@ -669,7 +669,7 @@ public class ProcessControllerTests {
         one.setSourcePackageId("A");
         var sd = new SentencetDetailType();
 
-        one.setSentencetDetail(Collections.singletonList(sd));
+        one.getSentencetDetail().add(sd);
         var af = new SentenceApplyToFileType();
         af.setAppearanceId("A");
         af.setFileNumber("A");
@@ -690,7 +690,7 @@ public class ProcessControllerTests {
         var dt = new DurationType();
         dt.setDurationLength("A");
         dt.setDurationUnitCd("A");
-        mt.setDuration(Collections.singletonList(dt));
+        mt.getDuration().add(dt);
 
         var nt = new NonMonetaryTermType();
         nt.setSentenceTypeCd("A");
@@ -706,12 +706,12 @@ public class ProcessControllerTests {
 
         var tt = new TermType();
         tt.setTermTypeCd("A");
-        tt.setDuration(Collections.singletonList(dt));
-        nt.setTerm(Collections.singletonList(tt));
+        tt.getDuration().add(dt);
+        nt.getTerm().add(tt);
 
-        sd.setNonMonetaryTerm(Collections.singletonList(nt));
-        sd.setMonetaryTerm(Collections.singletonList(mt));
-        sd.setApplyToFile(Collections.singletonList(af));
+        sd.getNonMonetaryTerm().add(nt);
+        sd.getMonetaryTerm().add(mt);
+        sd.getApplyToFile().add(af);
 
         req.setSentence(one);
 
@@ -766,9 +766,9 @@ public class ProcessControllerTests {
         bi.setBanCommentText("A");
         bi.setAcprId("A");
 
-        bt.setBanInfo(Collections.singletonList(bi));
-        bt.setApplyToFileBan(Collections.singletonList(af));
-        one.setBanDetail(Collections.singletonList(bt));
+        bt.getBanInfo().add(bi);
+        bt.getApplyToFileBan().add(af);
+        one.getBanDetail().add(bt);
         req.setBan(one);
 
         var out = new ProcessBanResponse();
@@ -804,14 +804,14 @@ public class ProcessControllerTests {
         af.setAppearanceId("A");
         af.setFileNumber("A");
         af.setFileCourtDivisionCd("A");
-        one.setApplyToFile(Collections.singletonList(af));
+        one.getApplyToFile().add(af);
 
         var nd = new NoteDetailType();
         nd.setNoteText("A");
         nd.setNoteSeqNo("A");
         nd.setNoteTypeCd("A");
 
-        one.setNoteDetail(Collections.singletonList(nd));
+        one.getNoteDetail().add(nd);
         req.setNote(one);
 
         var out = new ProcessNoteResponse();
@@ -854,14 +854,14 @@ public class ProcessControllerTests {
         var af = new ApplyToFileArraignmentType();
         af.setAppearanceId("A");
         af.setFileNumber("A");
-        ad.setApplyToFileArraignment(Collections.singletonList(af));
+        ad.getApplyToFileArraignment().add(af);
 
         var ae = new ArraignmentEventType();
         ae.setEventType("A");
         ae.setDetailText("A");
 
-        ad.setArraignmentEvent(Collections.singletonList(ae));
-        one.setArraignmentDetail(Collections.singletonList(ad));
+        ad.getArraignmentEvent().add(ae);
+        one.getArraignmentDetail().add(ad);
         req.setArraignment(one);
 
         var out = new ProcessArraignmentResponse();
@@ -905,8 +905,8 @@ public class ProcessControllerTests {
         af.setAppearanceId("A");
         af.setFileCourtDivisionCd("A");
         af.setFileNumber("A");
-        md.setApplyToFileMove(Collections.singletonList(af));
-        one.setMoveDetail(Collections.singletonList(md));
+        md.getApplyToFileMove().add(af);
+        one.getMoveDetail().add(md);
         req.setMove(one);
 
         var out = new ProcessMoveResponse();
@@ -949,7 +949,7 @@ public class ProcessControllerTests {
         fd.setSectionOtherStatuteId("A");
         fd.setSectionOther("A");
 
-        one.setFindingDetails(Collections.singletonList(fd));
+        one.getFindingDetails().add(fd);
 
         var af = new ApplyToFileFindingType();
         af.setAppearanceId("A");
@@ -957,7 +957,7 @@ public class ProcessControllerTests {
         af.setCountId("A");
         af.setCountNumber("A");
 
-        fd.setApplyToFileFinding(Collections.singletonList(af));
+        fd.getApplyToFileFinding().add(af);
 
         var out = new ProcessFindingResponse();
         out.setStatus("A");
@@ -1003,9 +1003,9 @@ public class ProcessControllerTests {
         af.setCountId("A");
         af.setCountNumber("A");
 
-        rd.setApplyToFileGenericResults(Collections.singletonList(af));
+        rd.getApplyToFileGenericResults().add(af);
 
-        one.setResultDetail(Collections.singletonList(rd));
+        one.getResultDetail().add(rd);
 
         req.setGenericResult(one);
 
@@ -1050,7 +1050,7 @@ public class ProcessControllerTests {
         af.setFileNumber("A");
         af.setAppearanceId("A");
 
-        cd.setApplyToFile(Collections.singletonList(af));
+        cd.getApplyToFile().add(af);
 
         var cm = new CivilPartyAppearanceMethodType();
         cm.setPartyName("A");
@@ -1063,18 +1063,18 @@ public class ProcessControllerTests {
         ct.setCounselFullName("A");
         ct.setPartyId("A");
         ct.setAttendanceMethodCd("A");
-        cm.setCounsel(Collections.singletonList(ct));
+        cm.getCounsel().add(ct);
 
         var rt = new RepresentativeType();
         rt.setRepresentativeFullName("A");
         rt.setAttendanceMethodCd("A");
         rt.setPartyId("A");
 
-        cm.setRepresentative(Collections.singletonList(rt));
+        cm.getRepresentative().add(rt);
 
-        cd.setPartyAppearanceMethod(Collections.singletonList(cm));
+        cd.getPartyAppearanceMethod().add(cm);
 
-        one.setCivilAppearanceMethodDetail(Collections.singletonList(cd));
+        one.getCivilAppearanceMethodDetail().add(cd);
         req.setCivilAppearanceMethod(one);
 
         var out = new ProcessCivilAppearanceMethodResponse();
@@ -1116,7 +1116,7 @@ public class ProcessControllerTests {
         var af = new ApplyToFileOrderType();
         af.setAppearanceId("A");
         af.setFileNumber("A");
-        od.setApplyToFile(Collections.singletonList(af));
+        od.getApplyToFile().add(af);
 
         var oe = new OrderEventType();
         oe.setOrderDetailText("A");
@@ -1127,9 +1127,9 @@ public class ProcessControllerTests {
         oe.setCustody("A");
         var oct = new OrderTypeCdType();
         oct.setOrderTypeCd("A");
-        oe.setOrderTypeCds(Collections.singletonList(oct));
+        oe.getOrderTypeCds().add(oct);
 
-        one.setOrderDetail(Collections.singletonList(od));
+        one.getOrderDetail().add(od);
         req.setOrder(one);
 
         var out = new ProcessOrderResponse();
@@ -1189,37 +1189,37 @@ public class ProcessControllerTests {
         af.setFileNumber("A");
         af.setDocumentId("A");
         af.setFileSeqNo("A");
-        co.setApplyToFile(Collections.singletonList(af));
+        co.getApplyToFile().add(af);
 
         var ot = new OrderTermType();
         ot.setOrderTermSeqNo("A");
         ot.setOrderTermEventTime(Instant.now());
         ot.setOrderTermText("A");
         ot.setOrderTermLogNoteText("A");
-        co.setOrderTerm(Collections.singletonList(ot));
+        co.getOrderTerm().add(ot);
 
         var op = new CivilOrderPartyType();
         op.setOrderRoleCd("A");
         op.setPartyFullName("A");
         op.setPartyId("A");
         op.setBirthDate(Instant.now());
-        co.setParty(Collections.singletonList(op));
+        co.getParty().add(op);
 
-        one.setCivilOrderDetail(Collections.singletonList(co));
+        one.getCivilOrderDetail().add(co);
 
         req.setCivilOrderResult(one);
 
         var ao = new AdminOrderDetailType();
-        ao.setApplyToFile(Collections.singletonList(af));
-        ao.setOrderTerm(Collections.singletonList(ot));
+        ao.getApplyToFile().add(af);
+        ao.getOrderTerm().add(ot);
 
-        one.setAdminOrderDetail(Collections.singletonList(ao));
+        one.getAdminOrderDetail().add(ao);
 
         var jc = new JudgeCommentType2();
         jc.setAppearanceId("A");
         jc.setJudgeCommentText("A");
 
-        one.setJudgeComment(Collections.singletonList(jc));
+        one.getJudgeComment().add(jc);
 
         var out = new ProcessCivilOrderResultResponse();
         out.setStatus("A");
@@ -1266,7 +1266,7 @@ public class ProcessControllerTests {
         af.setAppearanceId("A");
         af.setFileNumber("A");
 
-        ed.setApplyToFile(Collections.singletonList(af));
+        ed.getApplyToFile().add(af);
 
         var ei = new ExhibitItem();
         ei.setExhibitNo("A");
@@ -1274,8 +1274,8 @@ public class ProcessControllerTests {
         ei.setExhibitDescription("A");
         ei.setOwnedBy("A");
 
-        ed.setExhibitItem(Collections.singletonList(ei));
-        one.setExhibitDetail(Collections.singletonList(ed));
+        ed.getExhibitItem().add(ei);
+        one.getExhibitDetail().add(ed);
 
         req.setExhibitRequest(one);
 
@@ -1320,9 +1320,9 @@ public class ProcessControllerTests {
         af.setAppearanceId("A");
         af.setFileNumber("A");
 
-        sp.setApplyToFile(Collections.singletonList(af));
+        sp.getApplyToFile().add(af);
 
-        sp.setSpecialCourtCd(Collections.singletonList("A"));
+        sp.getSpecialCourtCd().add("A");
         req.setSpecialCourt(one);
         var out = new ProcessSpecialCourtResponse();
         out.setStatus("A");
